+++ src/main/java/racingcar/repository/PlayerResultDao.java
@@ -1,4 +1,4 @@
-package racingcar.dao;
+package racingcar.repository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
